Renamed new radio2 module as radio, and removed old code
