+++ src/monitor.py
@@ -9,8 +9,7 @@
 # However, it will log all messages from MiHome monitor, adaptor plus and house monitor
 # to a CSV log file, so could be the basis for a non-controlling energy logging app.
 
-from energenie import Registry, Devices, Messages, OpenThings
-from energenie import radio2 as radio
+from energenie import Registry, Devices, Messages, OpenThings, radio
 
 import Logger
