+++ src/switch.py
@@ -10,8 +10,7 @@
 # Don't expect this to be a good starting point for an application.
 # Consider waiting for me to finish developing the device object interface first.
 
-from energenie import Devices, Messages, Registry, OpenThings
-from energenie import radio2 as radio
+from energenie import Devices, Messages, Registry, OpenThings, radio
 from Timer import Timer
